refactor(annotate): extract shared helpers to _shared.py
_format_value, _ensure_renderer, _maybe_expand_limits, _expand_axis,
_format_column_value, _build_column_label_table were duplicated across
bar_values, point_values, box_stats, bar_custom. Factor them into
_shared.py; callers use maybe_expand_limits(axes_to_expand=...) so each
strategy decides which axes matter. bar_values keeps a _maybe_expand_limits
shim for bar_custom's import. No behavior change.
